refactor: Move typed object methods to common object
- Moved `getObjectEntries` to `TypedObject.entries`
- Moved `getObjectKeys` to `TypedObject.keys`
