+++ README.md
@@ -96,7 +96,6 @@ carry the overhead of checking argument types or handling type coercion.
 
 -   `entriesToKeyedItems(key: string, obj: object)`
 -   `filterLeavesByKey(key: string, obj: object, options)`
--   `getObjectKeys(obj: object | Array): string[]`
 -   `getOrDefault(obj: object, key: string, default: any)`
 -   `isObject(value: any): boolean`
 -   `mergeIf(conditional: any, :object)`
@@ -106,6 +105,8 @@ carry the overhead of checking argument types or handling type coercion.
 -   `omitUndefinedEntries(obj: object): object`
 -   `setValueByPath(path: string | string[], value: any, target: object, options?)`
 -   `swapKeysAndValues(obj: object): object`
+-   `TypedObject.entries(obj: object | Array | number | string): [key: string, value][]`
+-   `TypedObject.keys(obj: object | Array | number | string): string[]`
 
 ### String functions
