+++ src/functions/object/TypedObject/TypedObject.ts
@@ -1,0 +1,7 @@
+import { getObjectEntries } from './getObjectEntries';
+import { getObjectKeys } from './getObjectKeys';
+
+export const TypedObject = {
+  entries: getObjectEntries,
+  keys: getObjectKeys,
+};
