+++ src/functions/object/TypedObject/getObjectEntries.ts
