+++ src/functions/object/TypedObject/index.ts
@@ -1,0 +1,1 @@
+export { TypedObject } from './TypedObject';
