+++ src/functions/object/index.ts
@@ -1,6 +1,5 @@
 export { entriesToKeyedItems } from './entriesToKeyedItems';
 export { filterLeavesByKey } from './filterLeavesByKey';
-export { getObjectKeys } from './getObjectKeys';
 export { getOrDefault } from './getOrDefault';
 export { isObject } from './isObject';
 export { mergeIf } from './mergeIf';
@@ -10,3 +9,4 @@ export { omitFalsyEntries } from './omitFalsyEntries';
 export { omitUndefinedEntries } from './omitUndefinedEntries';
 export { setValueByPath } from './setValueByPath';
 export { swapKeysAndValues } from './swapKeysAndValues';
+export { TypedObject } from './TypedObject';
